Moving code to electron
Need to move to allow for console output from the developer console in electron to show up in the terminal on the electron run command in shell
